requirements.txt
pandas
requests
sqlalchemy
pymysql
openpyxl
python-dotenv
cryptography